Separate dequeuing "client" code into worker.py
Part of the problem with our original example is that
queueing/dequeueing both happen in the same process, so you get no real
benefit of delayed execution. You'd be waiting on the queued functions
in the main process regardless.

By moving them into a separate worker process, we can lay the ground
work for our multi-process app/worker future.

This current iteration has a major problem now, however. Our queue is
still an in-memory queue. That means that there is no way for our app
code and our worker process to share a queue, and therefore there's no
way for this iteration to work at all. The solution is to introduce a
"broker" so our app and our workers can share a queue. For this
"broker", we'll use Redis.