Introduce redis as a broker for our queue
Now we change the queue from an in-memory queue to being brokered by
Redis. This is useful for a number of reasons:

* If ether our app or worker processes crash, we don't lose the work
currently sitting in the queue
* Our app and worker processes can now read/write from the same queue

There is a known bug in this version, too, however. We're not
serializing the work to be done in any way when we hand it over to
Redis. Our in-memory queue, being a Python list, knew what to do when it
was given a Python object, but Redis doesn't know anything about Python
objects. Right now we're storing only the string representation of the
job, which is pretty useless since we won't be able to actually do the
work we're queueing. Luckily for us, Python has a serialization library
baked into its standard library...