Make worker more hands-off
Now our worker will run forever and just wait for work to be added onto
the queue.

The real magic here is in the change from using `rpop` to `brpop` when
popping an item from the queue. The use of `brpop` (blocking right pop)
keeps our worker from asking Redis for work as quickly as it can, and keeps
us from having to poll Redis for work.  `brpop` will try to pop something off
a list, but if the list is empty, it will patiently wait until there something
in the list it can pop off.

At this point, we're at a good spot with our basic queue. We can start a
worker and it'll happily just run and do work as it comes in. And our
app can queue work at its leisure.